# src/main/java/it/iet/interfaces/facade/dto/user/ChangeRoleDTO.java
package it.iet.interfaces.facade.dto.user;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class ChangeRoleDTO {
    
    String userMail;
    String role;
}